Little-endian. int16 messages. "OUTR" prefixes.

+++ macOS/LibOuterframeContent/OuterlayerHost.swift
@@ -137,7 +137,7 @@ class OuterlayerHost: NSObject {
         exit(0)
     }
 
-    func infraSocket(_ socket: InfraSocket, didReceiveMessageType typeRaw: UInt8, payload: Data) {
+    func infraSocket(_ socket: InfraSocket, didReceiveMessageType typeRaw: UInt16, payload: Data) {
         let message: BrowserToContentInfraMessage
         do {
             message = try BrowserToContentInfraMessage.decode(typeRaw: typeRaw, payload: payload)
